// fredob2b/src/app/services/api/fredob2b/models/shoping-card-detail-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ShopingCardDetailDto {
  createdBy?: number;
  createdDate?: string;
  description?: null | string;
  id?: number;
  imageName?: null | string;
  modifiedBy?: number;
  modifiedDate?: string;
  productId?: number;
  quantity?: number;
  taxLine?: number;
  totalLine?: number;
  unitPrice?: number;
}